added semester restiction in controller

+++ app/Http/Controllers/Admin/ProgramCourseController.php
@@ -16,6 +16,10 @@ class ProgramCourseController extends Controller
     {
         $this->authorize('update', $program);
 
+        if ($semester > $program->degree->semesters) {
+            return Redirect::back()->with('error', 'Semester is not available.');
+        }
+
         $courses =  Course::orderBy('course_name')
             ->get()
             ->transform(fn ($course) => [
@@ -41,18 +45,31 @@ class ProgramCourseController extends Controller
 
     public function store(Request $request, Program $program, $semester)
     {
-
         $this->authorize('update', $program);
 
+        $request->validate([
+            'courses' => 'required|array|min:1',
+        ]);
+
+        if ($semester > $program->degree->semesters) {
+            return Redirect::back()->withErrors(['The semester you selected is not available for this program.']);
+        }
+
+        //Our frontend form will send us all courses
         $courses = $request->courses;
 
+
+        //Filter where course belongs to program, and grab course ids
         $selected_courses = array_column(array_filter($courses, function ($course) {
             return $course['belongs_to_program'] == true;
         }), 'id');
 
 
+        //Detaching course where semester is given semester from program
         $program->courses()->wherePivot('semester', '=', $semester)->detach();
 
+
+        //Attach courses sent from frontend to program with given semester
         $program->courses()->attach($selected_courses, ['semester' => $semester]);
 
         return Redirect::route('admin.programs.edit', $program->slug)->with('success', 'Courses added successfully');
